+++ contracts/cw20-merkle-airdrop/src/lib.rs
@@ -7,3 +7,6 @@ pub mod msg;
 pub mod state;
 
 pub use crate::error::ContractError;
+
+#[cfg(test)]
+mod tests;
